list_all_music_albums class to list data
