+++ controllers/music.rb
@@ -1,0 +1,6 @@
+require_relative '../helpers/data_manager'
+class MusicAlbum < DataManager
+  def initialize
+    super('music')
+  end
+end
