+++ helpers/data_manager.rb
@@ -1,3 +1,4 @@
+require_relative 'file_manager'
 class DataManager
   def initialize(file_name)
     @file_storage = FileManager.new(file_name)
