+++ storage/music.json
@@ -1,1 +1,7 @@
-{"music" : []}
+[
+    {
+        "id": 252,
+        "public_date": "12/05/2023",
+        "on_spotify" : true
+    }
+]
